+++ PROYECTO/PaymentGateway/gateway/src/main/java/ec/edu/espe/gateway/comision/model/ComisionSegmentoPK.java
@@ -1,7 +1,6 @@
 package ec.edu.espe.gateway.comision.model;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 
@@ -12,15 +11,15 @@ public class ComisionSegmentoPK implements Serializable {
     @Column(name = "COD_COMISION", nullable = false)
     private Integer comision;
     @NotNull
-    @Column(name = "TRANSACCIONES_DESDE", precision = 9, nullable = false)
-    private BigDecimal transaccionesDesde;
+    @Column(name = "TRANSACCIONES_DESDE", nullable = false)
+    private Integer transaccionesDesde;
 
     // Empty constructor
     public ComisionSegmentoPK() {
     }
 
     // Constructor
-    public ComisionSegmentoPK(Integer comision, BigDecimal transaccionesDesde) {
+    public ComisionSegmentoPK(Integer comision, Integer transaccionesDesde) {
         this.comision = comision;
         this.transaccionesDesde = transaccionesDesde;
     }
@@ -34,12 +33,12 @@ public class ComisionSegmentoPK implements Serializable {
         this.comision = comision;
     }
 
-    public BigDecimal getTransaccionesDesde() {
+    public Integer getTransaccionesDesde() {
         return this.transaccionesDesde;
     }
 
-    public void setTransaccionesDesde(BigDecimal transaccionesDesde) {
-        this.transaccionesDesde = transaccionesDesde;
+    public void setTransaccionesDesde(Integer transaccionesDesde) {
+            this.transaccionesDesde = transaccionesDesde;
     }
 
     // ToString
